direwolf: fix do_configure error for ARM32 architecture
Error text looked like this:

| DEBUG: Executing shell function do_configure
[...]
| CMake system: Linux
| -- Target architecture: ARM
| CMake Error: TRY_RUN() invoked in cross-compiling mode, please set the following cache variables appropriately:
|    RUN_NEON (advanced)
| For details see [...]/tmp/work/cortexa7t2hf-neon-vfpv4-poky-linux-gnueabi/direwolf/1.6-r0/build/TryRunResults.cmake
[...]
| -- Configuring incomplete, errors occurred!
| See also "[...]/tmp/work/cortexa7t2hf-neon-vfpv4-poky-linux-gnueabi/direwolf/1.6-r0/build/CMakeFiles/CMakeOutput.log".
| See also "[...]/tmp/work/cortexa7t2hf-neon-vfpv4-poky-linux-gnueabi/direwolf/1.6-r0/build/CMakeFiles/CMakeError.log".
| WARNING: exit code 1 from a shell command.
ERROR: Task ([...]/meta-amateurradio/recipes-ham/direwolf/direwolf_1.6.bb:do_configure) failed with exit code '1'

Basically, when targeting ARM32 architectures, the cmake system
was checking the NEON support in the processor by attempting to
run a snippet of code[1] (which, when cross-compiling, does not make
sense), and so generating an error.

The fix involved passing the expected value of the RUN_NEON
variable to avoid performing the check.

[1] - https://github.com/wb2osz/direwolf/blob/1.6/cmake/modules/FindCPUflags.cmake#L357